+++ src/pl/plajer/villagedefense3/villagedefenseapi/StatsStorage.java
@@ -26,6 +26,7 @@ import java.util.TreeMap;
 import java.util.UUID;
 
 import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.craftbukkit.libs.jline.internal.Nullable;
 import org.bukkit.entity.Player;
 
 import pl.plajer.villagedefense3.Main;
@@ -100,6 +101,16 @@ public class StatsStorage {
 
     public String getName() {
       return name;
+    }
+
+    @Nullable
+    public static StatisticType getByName(String name) {
+      for (StatisticType s : values()) {
+        if (s.getName().equals(name)) {
+          return s;
+        }
+      }
+      return null;
     }
   }