+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added dynamic locale manager system - you can now get latest locales on demand from our repository
 * Added Romanian locale
 * Added 1.13.1 support
+* Fixed scoreboard color bugs (see https://i.imgur.com/kaZy5s2.png)
 
 ### 3.10.1 Release (13/17.08.2018)
 * Fixed NullPointerException in combust event (reported anonymously via Error service) (#1 error service report)